+++ src/application/Randomizer.h
@@ -46,7 +46,7 @@ public:
     void SetRingLink(bool ringLinkActive);
     void SetHardRingLink(bool hardRingLinkActive);
     void SetRingLoss(RingLoss ringLoss);
-    void SetSublevelChecks(bool sublevelChecks);
+    void SetSkyChaseChecks(bool skyChaseChecks);
     void SetBossChecks(bool bossChecks);
     void SetUnifyChaos4(bool unifyChaos4);
     void SetUnifyChaos6(bool unifyChaos6);
@@ -62,6 +62,9 @@ public:
     void OnPinballLifeCapsulesSet(bool pinballLifeCapsules);
     void ResetItems();
     void SetMissionMode(int missionModeEnabled);
+    void SetAutoStartMissions(int autoStartMissions);
+    void OnCheckVersion(int serverVersion);
+    void SetStartingCharacter(int startingCharacterIndex);
 
 private:
     bool AreLastStoryRequirementsCompleted() const;