+++ src/input/eventDetector/EventDetector.h
@@ -18,11 +18,8 @@ public:
     explicit EventDetector(Randomizer& randomizer);
     void OnPlayingFrame() const;
     void OnLevelEmblem(int character, int level, int mission);
-    void OnBossCompleted(int character, int level);
     void OnGenericEmblem(int index);
     void SetMultipleMissions(bool completeMultipleMissions);
-    void OnLevelCompleted(short character, short level);
-    void OnSublevelCompleted(short character, short level, int mission);
     LastStoryState lastStoryState = LastStoryNotStarted;
     bool completeMultipleLevelMissions = true;
     std::vector<LifeBoxLocationData> lifeCapsules;