+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.8.0] - 2024-08-?
+
+### Added
+
+- Level completion as a goal.
+- Complete all the action stages enabled to unlock the Perfect Chaos fight.
+- Added option to auto-start missions.
+- Added Angel Island area as a starting location.
+- Added Twinkle Park Lobby as a starting location.
+- Your starting character will now be selected on the character select screen when starting the game.
+- Added warning message for version mismatch when connecting to Archipelago.
+- Separated the sublevel checks options into two.
+    - One for Twinkle Circuit and Sand Hill and another for Sky Chase Act 1 and 2.
+- Added error messages when generating with really low settings. Now there is a minimum of 5 emblems for the goal.
+
+### Changed
+
+- Fixed normal/hard logic for missions not working correctly.
+- Added an 80% maximum for the emblem percentage goal.
+- Twinkle circuit is now in logic if you have access to the Twinkle Park Lobby.
+- Improved emblem requirements calculation.
+    - It should hopefully fix the issue with the wrong emblem percentage goal.
+- Fixed soft lock when losing any race as Tails.
+- Moved the City Hall emblem a little higher so you can get it when using the Cream the Rabbit mod.
+
 ## [v0.7.1] - 2024-08-22
 
 ### Added